Move classes to the correct packages
