+++ cypher/src/main/java/com/ldbc/impls/workloads/ldbc/snb/cypher/CypherDbConnectionState.java
@@ -2,7 +2,6 @@ package com.ldbc.impls.workloads.ldbc.snb.cypher;
 
 import com.ldbc.driver.DbException;
 import com.ldbc.impls.workloads.ldbc.snb.BaseDbConnectionState;
-import com.ldbc.impls.workloads.ldbc.snb.cypher.interactive.CypherQueryStore;
 import org.neo4j.driver.v1.AuthTokens;
 import org.neo4j.driver.v1.Driver;
 import org.neo4j.driver.v1.GraphDatabase;
