+++ postgres/src/main/java/com/ldbc/impls/workloads/ldbc/snb/postgres/operationhandlers/PostgresListOperationHandler.java
@@ -1,4 +1,4 @@
-package com.ldbc.impls.workloads.ldbc.snb.postgres.handlers;
+package com.ldbc.impls.workloads.ldbc.snb.postgres.operationhandlers;
 
 import com.ldbc.driver.DbException;
 import com.ldbc.driver.Operation;
@@ -10,39 +10,44 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.List;
 
-public abstract class PostgresSingletonOperationHandler<
-            OperationType extends Operation<OperationResult>,
+public abstract class PostgresListOperationHandler<
+            OperationType extends Operation<List<OperationResult>>,
             OperationResult
-        >
-        implements OperationHandler<OperationType, PostgresDbConnectionState> {
+        > implements OperationHandler<OperationType, PostgresDbConnectionState> {
 
     @Override
     public void executeOperation(OperationType operation, PostgresDbConnectionState state,
                                  ResultReporter resultReporter) throws DbException {
         Connection conn = state.getConnection();
-        OperationResult tuple = null;
+        List<OperationResult> results = new ArrayList<>();
         int resultCount = 0;
+        results.clear();
+
         String queryString = getQueryString(state, operation);
         try (final Statement stmt = conn.createStatement()) {
             state.logQuery(operation.getClass().getSimpleName(), queryString);
 
             ResultSet result = stmt.executeQuery(queryString);
-            if (result.next()) {
+            while (result.next()) {
                 resultCount++;
 
-                tuple = convertSingleResult(result);
+                OperationResult tuple = convertSingleResult(result);
                 if (state.isPrintResults())
                     System.out.println(tuple.toString());
+                results.add(tuple);
             }
-            stmt.close();
         } catch (Exception e) {
             throw new DbException(e);
         }
-        resultReporter.report(resultCount, tuple, operation);
+        resultReporter.report(resultCount, results, operation);
     }
 
     public abstract String getQueryString(PostgresDbConnectionState state, OperationType operation);
 
-    public abstract OperationResult convertSingleResult(ResultSet result) throws SQLException;
+    public OperationResult convertSingleResult(ResultSet result) throws SQLException {
+        throw new UnsupportedOperationException();
+    }
 }
