+++ postgres/src/main/java/com/ldbc/impls/workloads/ldbc/snb/postgres/operationhandlers/PostgresMultipleUpdateOperationHandler.java
@@ -1,4 +1,4 @@
-package com.ldbc.impls.workloads.ldbc.snb.postgres.handlers;
+package com.ldbc.impls.workloads.ldbc.snb.postgres.operationhandlers;
 
 import com.ldbc.driver.DbException;
 import com.ldbc.driver.Operation;
@@ -9,8 +9,9 @@ import com.ldbc.impls.workloads.ldbc.snb.postgres.PostgresDbConnectionState;
 
 import java.sql.Connection;
 import java.sql.Statement;
+import java.util.List;
 
-public abstract class PostgresUpdateOperationHandler<
+public abstract class PostgresMultipleUpdateOperationHandler<
             OperationType extends Operation<LdbcNoResult>
         > implements OperationHandler<OperationType, PostgresDbConnectionState> {
 
@@ -18,15 +19,19 @@ public abstract class PostgresUpdateOperationHandler<
     public void executeOperation(OperationType operation, PostgresDbConnectionState state,
                                  ResultReporter resultReporter) throws DbException {
         Connection conn = state.getConnection();
-        String queryString = getQueryString(state, operation);
-        try (final Statement stmt = conn.createStatement()) {
-            state.logQuery(operation.getClass().getSimpleName(), queryString);
-            stmt.execute(queryString);
+        try {
+            List<String> queryStrings = getQueryString(state, operation);
+            for (String queryString : queryStrings) {
+                Statement stmt = conn.createStatement();
+                state.logQuery(operation.getClass().getSimpleName(), queryString);
+                stmt.execute(queryString);
+                stmt.close();
+            }
         } catch (Exception e) {
             throw new DbException(e);
         }
         resultReporter.report(0, LdbcNoResult.INSTANCE, operation);
     }
 
-    public abstract String getQueryString(PostgresDbConnectionState state, OperationType operation);
+    public abstract List<String> getQueryString(PostgresDbConnectionState state, OperationType operation);
 }
