+++ postgres/src/main/java/com/ldbc/impls/workloads/ldbc/snb/postgres/operationhandlers/PostgresSingletonOperationHandler.java
@@ -1,4 +1,4 @@
-package com.ldbc.impls.workloads.ldbc.snb.postgres.handlers;
+package com.ldbc.impls.workloads.ldbc.snb.postgres.operationhandlers;
 
 import com.ldbc.driver.DbException;
 import com.ldbc.driver.Operation;
@@ -10,44 +10,39 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.List;
 
-public abstract class PostgresListOperationHandler<
-            OperationType extends Operation<List<OperationResult>>,
+public abstract class PostgresSingletonOperationHandler<
+            OperationType extends Operation<OperationResult>,
             OperationResult
-        > implements OperationHandler<OperationType, PostgresDbConnectionState> {
+        >
+        implements OperationHandler<OperationType, PostgresDbConnectionState> {
 
     @Override
     public void executeOperation(OperationType operation, PostgresDbConnectionState state,
                                  ResultReporter resultReporter) throws DbException {
         Connection conn = state.getConnection();
-        List<OperationResult> results = new ArrayList<>();
+        OperationResult tuple = null;
         int resultCount = 0;
-        results.clear();
-
         String queryString = getQueryString(state, operation);
         try (final Statement stmt = conn.createStatement()) {
             state.logQuery(operation.getClass().getSimpleName(), queryString);
 
             ResultSet result = stmt.executeQuery(queryString);
-            while (result.next()) {
+            if (result.next()) {
                 resultCount++;
 
-                OperationResult tuple = convertSingleResult(result);
+                tuple = convertSingleResult(result);
                 if (state.isPrintResults())
                     System.out.println(tuple.toString());
-                results.add(tuple);
             }
+            stmt.close();
         } catch (Exception e) {
             throw new DbException(e);
         }
-        resultReporter.report(resultCount, results, operation);
+        resultReporter.report(resultCount, tuple, operation);
     }
 
     public abstract String getQueryString(PostgresDbConnectionState state, OperationType operation);
 
-    public OperationResult convertSingleResult(ResultSet result) throws SQLException {
-        throw new UnsupportedOperationException();
-    }
+    public abstract OperationResult convertSingleResult(ResultSet result) throws SQLException;
 }
