+++ postgres/src/main/java/com/ldbc/impls/workloads/ldbc/snb/postgres/operationhandlers/PostgresUpdateOperationHandler.java
@@ -1,4 +1,4 @@
-package com.ldbc.impls.workloads.ldbc.snb.postgres.handlers;
+package com.ldbc.impls.workloads.ldbc.snb.postgres.operationhandlers;
 
 import com.ldbc.driver.DbException;
 import com.ldbc.driver.Operation;
@@ -9,9 +9,8 @@ import com.ldbc.impls.workloads.ldbc.snb.postgres.PostgresDbConnectionState;
 
 import java.sql.Connection;
 import java.sql.Statement;
-import java.util.List;
 
-public abstract class PostgresMultipleUpdateOperationHandler<
+public abstract class PostgresUpdateOperationHandler<
             OperationType extends Operation<LdbcNoResult>
         > implements OperationHandler<OperationType, PostgresDbConnectionState> {
 
@@ -19,19 +18,15 @@ public abstract class PostgresMultipleUpdateOperationHandler<
     public void executeOperation(OperationType operation, PostgresDbConnectionState state,
                                  ResultReporter resultReporter) throws DbException {
         Connection conn = state.getConnection();
-        try {
-            List<String> queryStrings = getQueryString(state, operation);
-            for (String queryString : queryStrings) {
-                Statement stmt = conn.createStatement();
-                state.logQuery(operation.getClass().getSimpleName(), queryString);
-                stmt.execute(queryString);
-                stmt.close();
-            }
+        String queryString = getQueryString(state, operation);
+        try (final Statement stmt = conn.createStatement()) {
+            state.logQuery(operation.getClass().getSimpleName(), queryString);
+            stmt.execute(queryString);
         } catch (Exception e) {
             throw new DbException(e);
         }
         resultReporter.report(0, LdbcNoResult.INSTANCE, operation);
     }
 
-    public abstract List<String> getQueryString(PostgresDbConnectionState state, OperationType operation);
+    public abstract String getQueryString(PostgresDbConnectionState state, OperationType operation);
 }
