+++ sparql/src/main/java/com/ldbc/impls/workloads/ldbc/snb/sparql/bi/SparqlBiDb.java
@@ -54,8 +54,9 @@ import com.ldbc.driver.workloads.ldbc.snb.bi.LdbcSnbBiQuery9RelatedForums;
 import com.ldbc.driver.workloads.ldbc.snb.bi.LdbcSnbBiQuery9RelatedForumsResult;
 import com.ldbc.impls.workloads.ldbc.snb.sparql.SparqlDb;
 import com.ldbc.impls.workloads.ldbc.snb.sparql.SparqlDbConnectionState;
-import com.ldbc.impls.workloads.ldbc.snb.sparql.SparqlListOperationHandler;
-import com.ldbc.impls.workloads.ldbc.snb.sparql.SparqlSingletonOperationHandler;
+import com.ldbc.impls.workloads.ldbc.snb.sparql.SparqlQueryStore;
+import com.ldbc.impls.workloads.ldbc.snb.sparql.operationhandlers.SparqlListOperationHandler;
+import com.ldbc.impls.workloads.ldbc.snb.sparql.operationhandlers.SparqlSingletonOperationHandler;
 import org.openrdf.query.BindingSet;
 
 import java.text.ParseException;
