+++ sparql/src/main/java/com/ldbc/impls/workloads/ldbc/snb/sparql/operationhandlers/SparqlSingletonOperationHandler.java
@@ -1,27 +1,23 @@
-package com.ldbc.impls.workloads.ldbc.snb.sparql;
+package com.ldbc.impls.workloads.ldbc.snb.sparql.operationhandlers;
 
 import com.ldbc.driver.DbException;
 import com.ldbc.driver.Operation;
 import com.ldbc.driver.OperationHandler;
 import com.ldbc.driver.ResultReporter;
+import com.ldbc.impls.workloads.ldbc.snb.sparql.SparqlDbConnectionState;
 import org.openrdf.query.BindingSet;
 import org.openrdf.query.TupleQuery;
 import org.openrdf.query.TupleQueryResult;
 
-import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class SparqlListOperationHandler<OperationType extends Operation<List<OperationResult>>, OperationResult>
+public abstract class SparqlSingletonOperationHandler<OperationType extends Operation<OperationResult>, OperationResult>
         implements OperationHandler<OperationType, SparqlDbConnectionState> {
 
     @Override
     public void executeOperation(OperationType operation, SparqlDbConnectionState state,
                                  ResultReporter resultReporter) throws DbException {
         try {
-            final List<OperationResult> results = new ArrayList<>();
+            OperationResult tuple = null;
             int resultCount = 0;
-            results.clear();
 
             final String queryString = getQueryString(state, operation);
             state.logQuery(operation.getClass().getSimpleName(), queryString);
@@ -29,19 +25,17 @@ public abstract class SparqlListOperationHandler<OperationType extends Operation
             TupleQuery tupleQuery = state.getRepository().getConnection().prepareTupleQuery(queryString);
             TupleQueryResult queryResults = tupleQuery.evaluate();
 
-            while (queryResults.hasNext()) {
+            if (queryResults.hasNext()) {
                 final BindingSet bs = queryResults.next();
                 resultCount++;
-                OperationResult tuple;
                 tuple = convertSingleResult(bs);
-                results.add(tuple);
                 if (state.isPrintResults()) {
                     System.out.println(tuple);
                 }
             }
             queryResults.close();
 
-            resultReporter.report(resultCount, results, operation);
+            resultReporter.report(resultCount, tuple, operation);
         } catch (Exception e) {
             throw new DbException(e);
         }
@@ -49,5 +43,5 @@ public abstract class SparqlListOperationHandler<OperationType extends Operation
 
     public abstract String getQueryString(SparqlDbConnectionState state, OperationType operation);
 
-    public abstract OperationResult convertSingleResult(BindingSet record) throws ParseException;
+    public abstract OperationResult convertSingleResult(BindingSet bs);
 }
